requirements.txt
# Core dependencies
kafka-python>=2.0.2
prometheus-client>=0.16.0
pyyaml>=6.0
pytest>=7.3.1

# Logging and monitoring
structlog>=23.1.0
python-json-logger>=2.0.7

# Kubernetes integration
kubernetes>=26.1.0

# Utilities
requests>=2.28.2
python-dotenv>=1.0.0

# Development tools
black>=23.3.0
flake8>=6.0.0
pytest-cov>=4.1.0 

# MongoDB
pymongo>=4.5.0

# Azure
azure-cosmos>=4.5.0
azure-storage-blob>=12.19.0
aiohttp>=3.8.0

# Time
python-dateutil>=2.8.2
arrow>=1.3.0